🐞 Corrige utm de compartilhamento thank_you mobile
